+++ cli/tests/integration/async_io.rs
@@ -1,12 +1,9 @@
-use crate::common::Test;
-use crate::common::TestResult;
-use hyper::Body;
-use hyper::Request;
-use hyper::Response;
-use hyper::StatusCode;
-use std::sync::atomic::AtomicUsize;
-use std::sync::atomic::Ordering;
-use std::sync::Arc;
+use crate::common::{Test, TestResult};
+use hyper::{body::HttpBody, Body, Request, Response, StatusCode};
+use std::sync::{
+    atomic::{AtomicUsize, Ordering},
+    Arc,
+};
 use tokio::sync::Barrier;
 
 // On Windows, streaming body backpressure doesn't seem to work as expected, either
@@ -22,11 +19,20 @@ async fn async_io_methods() -> TestResult {
     let req_count_1 = request_count.clone();
     let req_count_2 = request_count.clone();
     let req_count_3 = request_count.clone();
+    let req_count_4 = request_count.clone();
 
     let barrier = Arc::new(Barrier::new(3));
     let barrier_1 = barrier.clone();
     let barrier_2 = barrier.clone();
+    let sync_barrier = Arc::new(Barrier::new(2));
+    let sync_barrier_1 = sync_barrier.clone();
 
+    // We set up 4 async backends below, configured to test different
+    // combinations of async behavior from the guest.  The first three backends
+    // are things we are actually testing, and the fourth ("Semaphore") is just
+    // used as a synchronization mechanism. Each backend will receive 4 requests
+    // total and will behave differently depending on which request # it is
+    // processing.
     let test = Test::using_fixture("async_io.wasm")
         .async_backend("Simple", "/", None, move |req: Request<Body>| {
             assert_eq!(req.headers()["Host"], "simple.org");
@@ -34,25 +40,11 @@ async fn async_io_methods() -> TestResult {
             let barrier_1 = barrier_1.clone();
             Box::new(async move {
                 match req_count_1.load(Ordering::Relaxed) {
-                    0 => {
-                        barrier_1.wait().await;
-                        Response::builder()
-                            .status(StatusCode::OK)
-                            .body(Body::empty())
-                            .unwrap()
-                    }
                     1 => Response::builder()
                         .status(StatusCode::OK)
                         .body(Body::empty())
                         .unwrap(),
-                    2 => {
-                        barrier_1.wait().await;
-                        Response::builder()
-                            .status(StatusCode::OK)
-                            .body(Body::empty())
-                            .unwrap()
-                    }
-                    3 => {
+                    0 | 2 | 3 => {
                         barrier_1.wait().await;
                         Response::builder()
                             .status(StatusCode::OK)
@@ -69,21 +61,11 @@ async fn async_io_methods() -> TestResult {
             let req_count_2 = req_count_2.clone();
             Box::new(async move {
                 match req_count_2.load(Ordering::Relaxed) {
-                    0 => Response::builder()
-                        .header("Transfer-Encoding", "chunked")
-                        .status(StatusCode::OK)
-                        .body(Body::empty())
-                        .unwrap(),
-                    1 => Response::builder()
-                        .header("Transfer-Encoding", "chunked")
-                        .status(StatusCode::OK)
-                        .body(Body::empty())
-                        .unwrap(),
                     2 => Response::builder()
                         .status(StatusCode::OK)
                         .body(Body::empty())
                         .unwrap(),
-                    3 => Response::builder()
+                    0 | 1 | 3 => Response::builder()
                         .header("Transfer-Encoding", "chunked")
                         .status(StatusCode::OK)
                         .body(Body::empty())
@@ -97,36 +79,63 @@ async fn async_io_methods() -> TestResult {
             assert_eq!(req.headers()["Host"], "writebody.org");
             let req_count_3 = req_count_3.clone();
             let barrier_2 = barrier_2.clone();
+            let sync_barrier = sync_barrier.clone();
             Box::new(async move {
                 match req_count_3.load(Ordering::Relaxed) {
-                    0 => {
-                        barrier_2.wait().await;
-                        Response::builder()
-                            .status(StatusCode::OK)
-                            .body(Body::empty())
-                            .unwrap()
-                    }
-                    1 => {
-                        barrier_2.wait().await;
-                        Response::builder()
-                            .status(StatusCode::OK)
-                            .body(Body::empty())
-                            .unwrap()
-                    }
-                    2 => {
-                        barrier_2.wait().await;
-                        Response::builder()
-                            .status(StatusCode::OK)
-                            .body(Body::empty())
-                            .unwrap()
-                    }
                     3 => {
-                        let _body = hyper::body::to_bytes(req.into_body()).await;
+                        // Read at least 4MB and one 8K chunk from the request
+                        // to relieve back-pressure for the guest. These numbers
+                        // come from the amount of data that the guest writes to
+                        // the request body in test-fixtures/src/bin/async_io.rs
+                        let mut bod = req.into_body();
+                        let mut bytes_read = 0;
+                        while bytes_read < (4 * 1024 * 1024) + (8 * 1024) {
+                            if let Some(Ok(bytes)) = bod.data().await {
+                                bytes_read += bytes.len();
+                            }
+                        }
+
+                        // The guest will have another outstanding request to
+                        // the Semaphore backend below. Awaiting on the barrier
+                        // here will cause that request to return indicating to
+                        // the guest that we have read from the request body
+                        // and the write handle should be ready again.
+                        sync_barrier.wait().await;
+                        let _body = hyper::body::to_bytes(bod);
                         Response::builder()
                             .status(StatusCode::OK)
                             .body(Body::empty())
                             .unwrap()
                     }
+                    0 | 1 | 2 => {
+                        barrier_2.wait().await;
+                        Response::builder()
+                            .status(StatusCode::OK)
+                            .body(Body::empty())
+                            .unwrap()
+                    }
+                    _ => unreachable!(),
+                }
+            })
+        })
+        .await
+        .async_backend("Semaphore", "/", None, move |req: Request<Body>| {
+            assert_eq!(req.headers()["Host"], "writebody.org");
+            let req_count_4 = req_count_4.clone();
+            let sync_barrier_1 = sync_barrier_1.clone();
+            Box::new(async move {
+                match req_count_4.load(Ordering::Relaxed) {
+                    3 => {
+                        sync_barrier_1.wait().await;
+                        Response::builder()
+                            .status(StatusCode::OK)
+                            .body(Body::empty())
+                            .unwrap()
+                    }
+                    0 | 1 | 2 => Response::builder()
+                        .status(StatusCode::OK)
+                        .body(Body::empty())
+                        .unwrap(),
                     _ => unreachable!(),
                 }
             })