+++ cli/tests/integration/main.rs
@@ -1,10 +1,12 @@
 mod async_io;
 mod body;
+mod client_certs;
 mod common;
 mod dictionary_lookup;
 mod downstream_req;
 mod env_vars;
 mod geolocation_lookup;
+mod grpc;
 mod http_semantics;
 mod kv_store;
 mod logging;